+++ CHANGELOG.md
@@ -11,6 +11,10 @@
 
 - Upgrade wwdtm from 2.13.0 to 2.14.0
 
+### Development Changes
+
+- Add missing tests for Best Of, Repeat Best Of and Repeat Shows pages
+
 ## 6.2.0
 
 ### Application Changes